third_party/WebKit/LayoutTests/http/tests/security/mixedContent/laxChecking/websocket/insecure-websocket-in-secure-page-allowed-expected.txt

CONSOLE WARNING: line 9: The page at 'https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-successful-construction.html' was loaded over HTTPS, but ran insecure content from 'ws://127.0.0.1:8880/echo': this content should also be loaded over HTTPS.

CONSOLE WARNING: line 9: Connecting to a non-secure WebSocket server from a secure origin is deprecated.
This is a testharness.js-based test.
PASS Test that an insecure WebSocket connection is blocked when created in an HTTPS resource. 
Harness: the test ran to completion.
